// lib/api/generated/models/utils-payment-status.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Huerray Backend
 * Comprehensive API for connecting content creators with brands for marketing campaigns and gig opportunities
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@thrypes.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */




export const UtilsPaymentStatus = {
    PaymentStatusPending: 'pending',
    PaymentStatusProcessing: 'processing',
    PaymentStatusCompleted: 'completed',
    PaymentStatusFailed: 'failed',
    PaymentStatusCancelled: 'cancelled'
} as const;

export type UtilsPaymentStatus = typeof UtilsPaymentStatus[keyof typeof UtilsPaymentStatus];